Refactor, fix and create events
