+++ deck-gateway/src/main/kotlin/com/deck/gateway/entity/Messages.kt
@@ -2,21 +2,25 @@ package com.deck.gateway.entity
 
 import com.deck.common.entity.RawMessageContent
 import com.deck.common.util.GenericId
+import com.deck.common.util.OptionalProperty
 import com.deck.common.util.Timestamp
 import com.deck.common.util.UniqueId
 import kotlinx.serialization.Serializable
 
+/**
+ * Parameters [isSilent] [isPrivate] [repliesToIds] aren't present in webhook messages
+ */
 @Serializable
 data class RawPartialReceivedMessage(
     val id: UniqueId,
     val createdBy: GenericId,
     val content: RawMessageContent,
-    val repliesToIds: List<UniqueId>,
+    val repliesToIds: OptionalProperty<List<UniqueId>> = OptionalProperty.NotPresent,
     val repliesTo: UniqueId?,
     val type: String,
     val createdAt: Timestamp,
-    val isSilent: Boolean,
-    val isPrivate: Boolean
+    val isSilent: OptionalProperty<Boolean> = OptionalProperty.NotPresent,
+    val isPrivate: OptionalProperty<Boolean> = OptionalProperty.NotPresent
 )
 
 @Serializable
@@ -29,7 +33,14 @@ data class RawPartialRepliedMessage(
     val editedAt: Timestamp?,
     val deletedAt: Timestamp?,
     val channelId: UniqueId,
-    val webhookId: GenericId,
+    val webhookId: GenericId?,
     val isSilent: Boolean,
     val isPrivate: Boolean
+)
+
+@Serializable
+data class RawPartialDeletedMessage(
+    val id: UniqueId,
+    val deletedAt: Timestamp,
+    val createdAt: Timestamp
 )
