+++ deck-gateway/src/main/kotlin/com/deck/gateway/event/type/ChannelEvents.kt
@@ -1,0 +1,1 @@
+package com.deck.gateway.event.type
